Update "Linaro Toolchain" section of 'downloads'
Address https://linaro.zendesk.com/agent/tickets/1919

Update the "Linaro Toolchain" section of the www.linaro.org/downloads webpage to remove big-endian variants from the table of options, so as to not suggest that they are promoted to the same degree as the little-endian variants.  Also move linux and bare-metal targets into their own small tables.  Remove specific designation for windows host in the table entries, as the links were redundant with the link for the Linux host.  Add a release-notes link in each table entry.